move search functions to search.js in controllers folder
